v5_backup/backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
python-dotenv==1.0.1
openai==1.30.1
requests==2.32.3
pytest==8.2.2
gunicorn==22.0.0
beautifulsoup4==4.12.3
flask-bcrypt==1.0.1
flask-jwt-extended==4.7.1
google-auth==2.29.0
groq==0.8.0
razorpay

